Fixed recommender_test.py
In order to fix the test, it was necessary to update the names of the
Recommendation Strategies being imported while also changing the content of the
content based strategy from "full" to "mix".

It must be said that collaborative test were not fixed, because they are lacking
some data files.

Signed-off-by: Lucas Moura <lucas.moura128@gmail.com>